requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
gradio
scikit-learn
pandas
numpy
